refactor(core): refactor Amendtask import path in UC
